requirements.txt
transformers==4.46.3
sentence-transformers==3.3.1
streamlit
google-genai
chromadb
python-dotenv
firecrawl-py
pinecone
